Add files via upload
I have completed this assignment through UI i.e., through Pabbly account. As this to be done through the code, I have been stucked in between the assignment. As the deadline for this assignment have come ahead, please consider this document I have submitted which is done through an UI. If it is not taken under consideration please provide me some more time to complete the assignment through code. Hope you consider this assignment which is made by using UI.
Thank you.